fix(Main: typo) I thought we had an understanding that Motors was an instance.  Well, let me try again.
Was:
Ah!  Ok.  You didn't like me +ing them together.  I get it.  I'm sorry I yelled.

Was:
Ok, so a str.  Prove it.

Was:
What?!  Didn't like that either?  Still think a Motor is a str?!  How about THEM apples, huh??

Was:
Ok, so trying a new workaround.  I guess you can't iterate through a collection of Motors and
treat them as Motors.  Silly.  So I iterate through my known list of keys, and key into the list
of Motors.  Yeah, huh, ok?!  You try to mess with me, see how I take it!

Was:
Um...  Not sure why this broke.  The ingr_pumps is a collection of Motors.  Somehow, it thinks
a loop through all of them means each one is a str.  Hmm.  Ok.  So print it and see what it is.