Fix (minor): pump maintenance needed some custom stuff Finally redid the wiring for the pumps so it is now correct. Funny this is the pumps ran backwards\!\!  What's going on?
The pump wiring was correct.
The wiring on the jack at the Tiki Hut was correct.
The wiring on the jack at the box housing was correct.
The wiring on the Pi was (now) correct.

I realized that it was likely the wrong tubing that was the mistake\!
Reversed that and it works perfectly.

Also trimmed the tubing and recalibrated the Priming.
And replaced the Dark Run pump, sounding rough, and recalibrated it.