Kid drinks, fix lights on, BUG: kid drinks/RFIDs still linked At Obtanium, I added a feature that allowed a user to quickly iterate a new drink recipe.  Charlotte and Sullivan made some new kid drinks based on 12 new nonalcoholic ingredients. The new drinks they made are in KidDrinks.csv The drink recipe file is still hardcoded, so this file must be swapped out with TikiDrinks.csv.
BUG: The RFIDs are still hardcoded in DrinkBot.py  This means
the code won't be able to find the RFIDs for those drinks.
Fix: will need to strip the hardcoded RFIDs out of the code
and put them into the recipes file somehow.

Also in this git (sorry) I fixed the waiting for clear part
of mifareauth.py and now it works more like one would
expect -- you tell it to wait and it loops until it is clear
for X seconds.  if there's no wait, it loops until it reads
an RFID tag.

I also took out the ready_to_dispense function from the
initializing pumps function.  Now DrinkBot.py can turn on
and off the red dispense lights when the reader is clear.
Note that now the lights turn off when the drink is done
dispensing (red lights) and after dispensing, it needs to
be clear for 5 seconds.

Feature add: it would be better if the waiting can be done
while the drink is dispensing.  See if this is logaically
easy to add.